fix(ci): restrict Python versions in Python CI (#67)
Builds are currently failing because the maturin action is attempting to
build the wheel for every Python interpreter in the linux job. This
includes an alpha version of 3.13, but since 0.20.3 pyo3 will refuse to
build for that version of Python because forward-compatibility with
future 3.13 versions isn't guaranteed. We've also ended up with some
3.13 versions of augurs on PyPI :(

This commit restricts the versions to 3.7 - 3.12, the currently
supported versions of Python.

In future it's possible that maturin will allow us to only include
non-alpha/beta versions of Python with the '--find-interpreter' flag,
but until then we'll need to update them manually.

Relevant issue: https://github.com/PyO3/maturin/issues/1960